Switching to having an LEDController with separate LEDSubsystems
Having to define an enum per LED is a little gnarly, but we have tests
that verify they are defined correctly and it's pretty easy to generate
new ones.

The big benefit here is that different things can write to different
sections and as long as there isn't any overlap, they can all run
concurrently.