Don't warn about joysticks if we're in a simulation
It's very noisy. We could disable it on real robots too and it would
still work in competition, but that may lead someone astray if they're
unaware of that message being disabled. At least in simulation, it's
somewhat expected that your setup may at some point preclude an active
controller and it's not the end of the world.